Created a basic login function for login formula
When the user enters email and password and logs in, values from the fields are stored in localStorage.

As test, the app retrieves the data from localStorage and adds the values from the fields to console.

Prevented default formula behaviour.

the login button does nothing if the email and password fields are empty.