Login page
Added function to find esisting users.

When login in, the page can detect exsisting users. The login page uses localStorage to store a list of users

if it can find an exsisting user from localStorage, it will print out the users email on the page

if not, the page will tell it couldn't find the user